+++ src/__tests__/Call.test.ts
@@ -555,8 +555,8 @@ describe('Call class', () => {
 
     describe('.sendMesssage', () => {
       const content = { key1: 'hello world' };
-      const contentType = CallMessage.ContentType.ApplicationJson;
-      const messageType = CallMessage.MessageType.UserDefinedMessage;
+      const contentType = 'application/json';
+      const messageType = 'user-defined-message';
 
       it('invokes the native module', async () => {
         const message = new CallMessage({
@@ -585,6 +585,7 @@ describe('Call class', () => {
           contentType,
           messageType,
         });
+
         const sendMessagePromise = new Call(createNativeCallInfo()).sendMessage(
           message
         );