+++ CHANGELOG.md
@@ -1,12 +1,47 @@
-1.0.1 (In Development)
-====================
+1.1.0 (In Development)
+======================
 
 ## Features
+
 ### Platform Specific Features
 
 #### Android
+
 - Added support for Android 34
 
+## Fixes
+
+### Platform Specific Fixes
+
+#### iOS
+
+- Fixed Call Messages not being built with the passed `contentType` or `messageType`.
+
+## Changes
+
+### Call Message Events
+
+The Call Message Events feature, which was originally released under Beta in `1.0.0` of the SDK, is now promoted to Generally Available (GA).
+This release includes the following changes:
+
+- **(Breaking)** Removed `CallMessage.MessageType` and `CallMessage.ContentType` enumerations and types.
+  Instead, those types have been replaced by `string`.
+
+The following is an example of the updated API:
+```ts
+const call = await voice.connect(...);
+call.sendMessage(new CallMessage({
+  content: { foo: 'bar' },
+  contentType: 'application/json',
+  messageType: 'user-defined-message',
+}));
+```
+
+- Added new error codes. See the following table for details:
+  | Error Code | Description |
+  | --- | --- |
+  | 31210 | Raised when a Call Message is sent with an invalid message type. |
+  | 31211 | Raised when a Call Message is sent when the call is not yet ready to send messages. This can typically happen when the Call/CallInvite is not yet in a ringing state. |
 
 1.0.0 (Mar 25, 2024)
 ====================